Refactor MelAudioObfuscationEnv
